Update to new release of temperature-monitor
This has required some packaging changes:

It's now using a makefile, and inserting version information at build
time. This requires the use of pkg-info.mk to obtain the upstream
version from the package so that it can be passed to the build process.

It's also now vendoring dependencies into the vendor directory (using
the Go 1.5 vendor experiment), which therefore requires us to cd into
the GOPATH before building so that go can locate these dependencies.